test: add e2e tests on skip containers and token expiration
